Refactor (APIMF-3491): big step -> extracted property parsing
